fix(kernel): P2.1 PageQuery.getOffset() 返回类型改为 long 避免大偏移量溢出

+++ xcms-backend/xcms-shared-kernel/src/main/java/com/df4j/xctec/xcms/kernel/common/PageQuery.java
@@ -13,7 +13,7 @@ public class PageQuery {
     private String sortBy;
     private String sortOrder = "ASC";
 
-    public int getOffset() {
-        return (page - 1) * size;
+    public long getOffset() {
+        return (long) (page - 1) * size;
     }
 }
